src/main/java/com/example/cinema/service/MembershipService.java
package com.example.cinema.service;

import com.example.cinema.exceptions.NotFoundException;
import com.example.cinema.model.Membership;
import com.example.cinema.repository.MembershipRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class MembershipService {
    private final MembershipRepository membershipRepository;

    public MembershipService(MembershipRepository membershipRepository) {
        this.membershipRepository = membershipRepository;
    }

    public List<Membership> getAllMemberships() {
        return membershipRepository.findAll();
    }
}
